event-lifecycle-service/requirements.txt
# requirements.txt

# FastAPI framework and server
python-dotenv
fastapi
uvicorn[standard]

# Database and ORM
sqlalchemy
alembic
psycopg2-binary
sqlalchemy-utils

# Data validation and settings management
pydantic
pydantic-settings

# Security and JWT
python-jose[cryptography]
passlib[bcrypt]
PyJWT
slowapi  # Rate limiting for FastAPI

# For inter-service communication
redis
httpx
kafka-python-ng # <-- The modern, correct Kafka library
boto3
resend

# For background tasks
celery
pdf2image
apscheduler  # Scheduler for waitlist background tasks

# Payment processing
stripe

# Analytics and reporting
pandas
openpyxl  # For Excel export
scipy  # For A/B test statistical significance

# GraphQL and Federation
strawberry-graphql[fastapi]

# --- TESTING ---
pytest
